add cli commands for generating and reindexing table docs

+++ modules/langchain/index/create_index.py
@@ -8,11 +8,12 @@ from langchain.embeddings import HuggingFaceEmbeddings
 from langchain.indexes import VectorstoreIndexCreator
 
 from modules.config import config
-from modules.langchain.index import generate_table_documents
+from .generate_table_documents import generate_table_documents
 from .heavydb_metadata_index import HeavyDBMetadataIndex
 
 huggingface_model_name = config.huggingface_embed_model
 metadata_index_dir = config.metadata_index_dir
+table_documents_dir = config.table_documents_dir
 
 
 def read_table_documents(folder_path: str, include: Optional[list[str]] = None) -> Iterator[Document]:
@@ -37,12 +38,15 @@ def update_tables_in_index(
     index_creator: VectorstoreIndexCreator, vectorstore: Chroma, folder_path: str, tables_to_update: list[str]
 ):
     print("New table summaries will replace existing summaries in the index (if any).")
+    print("Reading new table documents...")
+    docs = list(read_table_documents(folder_path, include=tables_to_update))
+    if len(docs) != len(tables_to_update):
+        raise ValueError("Not all provided tables have documents. Aborting.")
+    print("Deleting existing table documents from index...")
     if len(tables_to_update) == 1:
         vectorstore._collection.delete(where={"source": tables_to_update[0]})
     else:
         vectorstore._collection.delete(where={"$or": [{"source": table_name} for table_name in tables_to_update]})
-    print("Reading new table documents...")
-    docs = list(read_table_documents(folder_path, include=tables_to_update))
     print("Splitting documents...")
     sub_docs = index_creator.text_splitter.split_documents(docs)
     print(f"Indexing documents with {huggingface_model_name}...")
@@ -62,7 +66,7 @@ def get_vectorstore_index_creator(persist_directory: str) -> VectorstoreIndexCre
     )
 
 
-def create_index_if_nonexistent(folder_path: str = "table_documents") -> HeavyDBMetadataIndex:
+def create_index_if_nonexistent() -> HeavyDBMetadataIndex:
     """
     Create a new vector store index if it does not exist, otherwise return the existing index.
 
@@ -70,7 +74,7 @@ def create_index_if_nonexistent(folder_path: str = "table_documents") -> HeavyDB
     :param persist_directory: Path to the directory where the vector store index should be persisted.
     :return: A HeavyDBMetadataIndex instance containing the vector store index.
     """
-    tables_with_new_summaries = generate_table_documents(folder_path)
+    tables_with_new_summaries = generate_table_documents(table_documents_dir)
 
     index_creator = get_vectorstore_index_creator(metadata_index_dir)
 
@@ -79,11 +83,11 @@ def create_index_if_nonexistent(folder_path: str = "table_documents") -> HeavyDB
         print("Index already exists. Returning existing index.")
         vectorstore: Chroma = Chroma(embedding_function=index_creator.embedding, persist_directory=metadata_index_dir)
         if len(tables_with_new_summaries) > 0:
-            update_tables_in_index(index_creator, vectorstore, folder_path, tables_with_new_summaries)
+            update_tables_in_index(index_creator, vectorstore, table_documents_dir, tables_with_new_summaries)
     else:
         print("Index does not exist. Creating new index.")
         print("Reading table documents...")
-        docs = read_table_documents(folder_path)
+        docs = read_table_documents(table_documents_dir)
 
         print("Splitting documents...")
         sub_docs = index_creator.text_splitter.split_documents(list(docs))
